Have smartcharts retrieve time from SmartTrader 2.0 instead of making server requests
