+++ src/javascript/_common/base/socket_base.js
@@ -5,6 +5,7 @@ const State            = require('../storage').State;
 const cloneObject      = require('../utility').cloneObject;
 const getPropertyValue = require('../utility').getPropertyValue;
 const isEmptyObject    = require('../utility').isEmptyObject;
+const PromiseClass     = require('../utility').PromiseClass;
 const getAppId         = require('../../config').getAppId;
 const getSocketURL     = require('../../config').getSocketURL;
 
@@ -291,14 +292,5 @@ const BinarySocketBase = (() => {
         removeOnDisconnect: () => { delete config.onDisconnect; },
     };
 })();
-
-class PromiseClass {
-    constructor() {
-        this.promise = new Promise((resolve, reject) => {
-            this.reject  = reject;
-            this.resolve = resolve;
-        });
-    }
-}
 
 module.exports = BinarySocketBase;
